Create better_snake?.c
A refactored version of `snake.c`, currently untested. `better_snake?.c` focuses on modularity and I will compare it to a version that will be focused on reduced function calls. Eventually, I plan to release the assembly code for both of these, as well as hopefully make a version with a bot snake that uses a path finding algorithm to get the highest score possible. 

I also have a version of the game that allows different difficulty levels to be selected, and I will release it sometime when I can clean it up.